Use "no feature qw(indirect)" to remove unwanted parentheses
It always bothered me that we often use ok() and is() without
parentheses but in some situations they couldn't be used
because of indirect object syntax (see `perldoc perlobj`).

With the pragma "indirect" disabled by default the
parentheses can be removed where we actually intended not to
use them.

Find candidates with:

    ack '^\w*\w+\(' t/